pyproject.toml
[tool.poetry]
name = "rubix-service"
version = "1.5.8"
description = "Install service via systemd"
authors = ["NubeIO"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.25.1"
netifaces = "^0.10.9"
gunicorn = "^20.0.4"
click = "^7.1.2"
gevent = [
    { version = "20.12.1", markers = "platform_machine == 'x86_64'" },
    { git = "https://github.com/gevent/gevent.git", tag = "20.12.1", markers = "platform_machine != 'x86_64'" }
]
packaging = "^20.8"
PyJWT = "^2.0.1"
rubix-registry = { git = "https://github.com/NubeIO/rubix-registry", rev = "v1.0.0" }
rubix-http = { git = "https://github.com/NubeIO/rubix-http", rev = "v1.0.2" }
mqtt-rest-bridge = { git = "https://github.com/NubeIO/mqtt-rest-bridge", rev = "v1.2.0" }

[tool.poetry.dev-dependencies]
pyinstaller = "^4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
